requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
demoji==1.1.0
et-xmlfile==1.1.0
exceptiongroup==1.0.4
h11==0.14.0
idna==3.4
Janome==0.4.2
lxml==4.9.1
numpy==1.23.5
oauthlib==3.2.2
openai==0.25.0
openpyxl==3.0.10
outcome==1.2.0
pandas==1.5.2
pandas-stubs==1.5.2.221124
PySocks==1.7.1
python-dateutil==2.8.2
pytrends==4.8.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.7.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
# termextract==0.12b
termextract @ http://gensen.dl.itc.u-tokyo.ac.jp/soft/pytermextract-0_02.zip
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tweepy==4.12.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.13
wsproto==1.2.0
